+++ socket.h
@@ -38,8 +38,8 @@ void make_dhcp_packet(struct udp_dhcp_packet *packet,
 		      struct in_addr source, struct in_addr dest);
 
 int open_socket (interface_t *iface, bool arp);
-int send_packet (const interface_t *iface, int type,
-		 const unsigned char *data, int len);
-int get_packet (const interface_t *iface, unsigned char *data,
-		unsigned char *buffer, int *buffer_len, int *buffer_pos);
+ssize_t send_packet (const interface_t *iface, int type,
+		     const unsigned char *data, int len);
+ssize_t get_packet (const interface_t *iface, unsigned char *data,
+		    unsigned char *buffer, int *buffer_len, int *buffer_pos);
 #endif
